Refactor of code to break up dart code, per pre demo PR
