+++ careconnect2025/frontend/lib/features/tasks/presentation/calendar_assisiant.dart
@@ -5,8 +5,10 @@
 import 'dart:collection';
 import 'dart:convert';
 
-import 'package:care_connect_app/features/notifications/models/scheduled_notification_model.dart';
 import 'package:care_connect_app/features/tasks/models/task_model.dart';
+import 'package:care_connect_app/features/tasks/utils/recurrence_utils.dart';
+import 'package:care_connect_app/features/tasks/utils/task_type_utils.dart';
+import 'package:care_connect_app/features/tasks/utils/task_utils.dart';
 import 'package:care_connect_app/providers/user_provider.dart';
 import 'package:care_connect_app/services/api_service.dart';
 import 'package:care_connect_app/widgets/app_bar_helper.dart';
@@ -15,383 +17,13 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:table_calendar/table_calendar.dart';
 
-// =============================
-// TaskCopyWith
-// =============================
-extension TaskCopyWith on Task {
-  Task copyWith({
-    int? id,
-    String? name,
-    String? description,
-    DateTime? date,
-    TimeOfDay? timeOfDay,
-    int? userId,
-    bool? isComplete,
-    String? frequency,
-    int? interval,
-    int? count,
-    List<bool>? daysOfWeek,
-    String? taskType,
-  }) {
-    return Task(
-      id: id ?? this.id,
-      name: name ?? this.name,
-      description: description ?? this.description,
-      date: date ?? this.date,
-      timeOfDay: timeOfDay ?? this.timeOfDay,
-      userId: userId ?? this.userId,
-      isComplete: isComplete ?? this.isComplete,
-      notifications: this.notifications, // keep existing notifications
-      frequency: frequency ?? this.frequency,
-      interval: interval ?? this.interval,
-      count: count ?? this.count,
-      daysOfWeek: daysOfWeek ?? this.daysOfWeek,
-      taskType: taskType ?? this.taskType,
-    );
-  }
-}
-
-/// Utilities for task normalization & grouping
-class TaskUtils {
-  static Map<String, dynamic> normalizeTaskMap(Map<String, dynamic> map) {
-    if (map['date'] is String) {
-      final d = map['date'];
-      if (!d.contains('T')) map['date'] = d.replaceFirst(' ', 'T');
-    }
-    final dow = map['daysOfWeek'];
-    if (dow is String) {
-      try {
-        map['daysOfWeek'] = List<bool>.from(jsonDecode(dow));
-      } catch (_) {
-        map['daysOfWeek'] = [];
-      }
-    }
-    if (map['isComplete'] == null && map['completed'] != null) {
-      map['isComplete'] = map['completed'];
-    }
-    return map;
-  }
-
-  static LinkedHashMap<DateTime, List<Task>> groupTasksByDate(
-    List<Task> tasks,
-  ) {
-    final grouped = LinkedHashMap<DateTime, List<Task>>(
-      equals: isSameDay,
-      hashCode: (date) => date.day * 1000000 + date.month * 10000 + date.year,
-    );
-
-    for (final task in tasks) {
-      final key = DateTime(task.date.year, task.date.month, task.date.day);
-
-      // Deduplicate by (task.id + task.date)
-      final existing = grouped.putIfAbsent(key, () => []);
-      if (!existing.any((t) => t.id == task.id && t.date == task.date)) {
-        existing.add(task);
-      }
-    }
-    return grouped;
-  }
-
-  /// Normalize a date to local midnight (drops hours/minutes/seconds).
-  static DateTime normalizeDate(DateTime d) {
-    return DateTime(d.year, d.month, d.day);
-  }
-}
-
-// =============================
-// RecurrenceUtils
-// =============================
-/// Utilities for building recurring tasks consistently
-class RecurrenceUtils {
-  static Task buildTask({
-    required Task baseTask,
-    bool? isRecurring,
-    String? recurrenceType, // "Daily", "Weekly", "Monthly", "Yearly"
-    List<bool>? daysOfWeek,
-    int? interval,
-    int? count,
-    DateTime? startDate,
-    DateTime? endDate,
-    int? dayOfMonth,
-  }) {
-    String? frequency;
-    int? intervalToSend = interval;
-    int? countToSend = count;
-
-    // Preserve original task type safely
-    final normalizedTaskType = baseTask.taskType?.toLowerCase();
-
-    // Default effective start date
-    DateTime effectiveDate = (startDate ?? baseTask.date).toLocal();
-    // Normalize end date if provided
-    DateTime? normalizedEndDate = endDate != null
-        ? TaskUtils.normalizeDate(endDate)
-        : null;
-
-    if (isRecurring == true && recurrenceType != null) {
-      switch (recurrenceType.toLowerCase()) {
-        case "daily":
-          frequency = "daily";
-          intervalToSend = (intervalToSend == null || intervalToSend < 1)
-              ? 1
-              : intervalToSend;
-
-          if (normalizedEndDate != null) {
-            int occurrences = 0;
-            DateTime cursor = effectiveDate;
-            while (!cursor.isAfter(normalizedEndDate)) {
-              occurrences++;
-              cursor = cursor.add(Duration(days: intervalToSend));
-            }
-            countToSend = occurrences;
-          } else {
-            countToSend ??= 30; // fallback: 30 days if no end date
-          }
-          break;
-
-        case "weekly":
-          frequency = "weekly";
-          intervalToSend = (intervalToSend == null || intervalToSend < 1)
-              ? 1
-              : intervalToSend;
-
-          if (normalizedEndDate != null) {
-            int occurrences = 0;
-            DateTime cursor = effectiveDate;
-            // walk week-by-week
-            while (!cursor.isAfter(normalizedEndDate)) {
-              for (int i = 0; i < 7; i++) {
-                if (daysOfWeek != null &&
-                    i < daysOfWeek.length &&
-                    daysOfWeek[i]) {
-                  // i=0 is Sunday → Dart DateTime.weekday is Mon=1…Sun=7
-                  int dartDow = (i == 0 ? DateTime.sunday : i + 1);
-                  DateTime candidate = cursor.subtract(
-                    Duration(days: cursor.weekday - dartDow),
-                  );
-
-                  if (!candidate.isBefore(effectiveDate) &&
-                      !candidate.isAfter(normalizedEndDate)) {
-                    occurrences++;
-                  }
-                }
-              }
-              cursor = cursor.add(Duration(days: 7 * intervalToSend));
-            }
-            countToSend = occurrences;
-          } else {
-            countToSend ??= 4; // default 4 weeks if no end date
-          }
-          break;
-
-        case "monthly":
-          frequency = "monthly";
-          intervalToSend = (intervalToSend == null || intervalToSend < 1)
-              ? 1
-              : intervalToSend;
-
-          int dom = dayOfMonth ?? effectiveDate.day;
-
-          // Adjust the first occurrence so it’s not before the startDate
-          final daysInStartMonth = DateUtils.getDaysInMonth(
-            effectiveDate.year,
-            effectiveDate.month,
-          );
-          dom = dom.clamp(1, daysInStartMonth);
-          DateTime firstCandidate = DateTime(
-            effectiveDate.year,
-            effectiveDate.month,
-            dom,
-          );
-
-          if (firstCandidate.isBefore(effectiveDate)) {
-            // push to next month
-            final nextMonth = effectiveDate.month + 1;
-            final nextYear = effectiveDate.year + ((nextMonth - 1) ~/ 12);
-            final adjustedMonth = ((nextMonth - 1) % 12) + 1;
-            final daysInNextMonth = DateUtils.getDaysInMonth(
-              nextYear,
-              adjustedMonth,
-            );
-            final nextDom = dom.clamp(1, daysInNextMonth);
-            firstCandidate = DateTime(nextYear, adjustedMonth, nextDom);
-          }
-
-          if (normalizedEndDate != null) {
-            int occurrences = 0;
-            DateTime cursor = firstCandidate;
-            while (!cursor.isAfter(normalizedEndDate)) {
-              occurrences++;
-              final nextMonth = cursor.month + intervalToSend;
-              final nextYear = cursor.year + ((nextMonth - 1) ~/ 12);
-              final adjustedMonth = ((nextMonth - 1) % 12) + 1;
-              final daysInNextMonth = DateUtils.getDaysInMonth(
-                nextYear,
-                adjustedMonth,
-              );
-              final nextDom = dom.clamp(1, daysInNextMonth);
-              cursor = DateTime(nextYear, adjustedMonth, nextDom);
-            }
-            countToSend = occurrences;
-            effectiveDate =
-                firstCandidate; // anchor to correct first occurrence
-          } else {
-            countToSend ??= 12; // fallback: 12 months
-            effectiveDate = firstCandidate;
-          }
-          break;
-
-        case "yearly":
-          frequency = "yearly";
-          intervalToSend = (intervalToSend == null || intervalToSend < 1)
-              ? 1
-              : intervalToSend;
-
-          if (normalizedEndDate != null) {
-            int occurrences = 0;
-            DateTime cursor = effectiveDate;
-            while (!cursor.isAfter(normalizedEndDate)) {
-              occurrences++;
-              cursor = DateTime(
-                cursor.year + intervalToSend,
-                cursor.month,
-                cursor.day,
-              );
-            }
-            countToSend = occurrences;
-          } else {
-            countToSend ??= 5; // default: 5 years
-          }
-          break;
-      }
-    }
-    return baseTask.copyWith(
-      date: effectiveDate,
-      frequency: frequency,
-      interval: intervalToSend,
-      count: countToSend,
-      daysOfWeek: daysOfWeek,
-    );
-  }
-
-  static int calculateCount({
-    required DateTime startDate,
-    required DateTime endDate,
-    required String frequency, // "daily", "weekly", "monthly", "yearly"
-    required int interval,
-    List<bool>? daysOfWeek, // for weekly: 7-length list Sunday..Saturday
-  }) {
-    switch (frequency.toLowerCase()) {
-      case "daily":
-        final diffDays = endDate.difference(startDate).inDays;
-        return (diffDays ~/ interval) + 1;
-
-      case "weekly":
-        if (daysOfWeek == null || daysOfWeek.length != 7) {
-          final diffWeeks = endDate.difference(startDate).inDays ~/ 7;
-          return (diffWeeks ~/ interval) + 1;
-        }
-        int count = 0;
-        DateTime cursor = startDate;
-        while (!cursor.isAfter(endDate)) {
-          if (daysOfWeek[cursor.weekday % 7]) {
-            count++;
-          }
-          cursor = cursor.add(const Duration(days: 1));
-        }
-        return count;
-
-      case "monthly":
-        int months =
-            (endDate.year - startDate.year) * 12 +
-            (endDate.month - startDate.month);
-        return (months ~/ interval) + 1;
-
-      case "yearly":
-        int years = endDate.year - startDate.year;
-        return (years ~/ interval) + 1;
-
-      default:
-        return 1;
-    }
-  }
-
-  static DateTime calculateEndDate({
-    required DateTime startDate,
-    required String frequency,
-    required int interval,
-    required int count,
-    List<bool>? daysOfWeek,
-  }) {
-    switch (frequency.toLowerCase()) {
-      case "daily":
-        return startDate.add(Duration(days: (count - 1) * interval));
-      case "weekly":
-        return startDate.add(Duration(days: (count - 1) * 7 * interval));
-      case "monthly":
-        return DateTime(
-          startDate.year,
-          startDate.month + (count - 1) * interval,
-          startDate.day,
-        );
-      case "yearly":
-        return DateTime(
-          startDate.year + (count - 1) * interval,
-          startDate.month,
-          startDate.day,
-        );
-      default:
-        return startDate;
-    }
-  }
-}
-
-// =============================
-// TaskTypeUtils.dart
-// =============================
-///Moving task Types into utility class for use across code base
-class TaskTypeUtils {
-  static final Map<String, Color> taskTypeColors = {
-    'medication': Colors.red,
-    'appointment': Colors.blue,
-    'exercise': Colors.green,
-    'general': Colors.deepOrange,
-    'lab': Colors.purple,
-    'pharmacy': Colors.teal,
-  };
-
-  /// Map used to change icon used for certain types of task
-  static final Map<String, IconData> taskTypeIcons = {
-    'medication': Icons.medication,
-    'appointment': Icons.event,
-    'exercise': Icons.fitness_center,
-    'general': Icons.task,
-    'lab': Icons.science,
-    'pharmacy': Icons.local_pharmacy,
-  };
-
-  /// Get a color safely with fallback
-  static Color getColor(String? type) {
-    if (type == null) return Colors.deepOrange;
-    return taskTypeColors[type.toLowerCase()] ?? Colors.deepOrange;
-  }
-
-  /// Get icon safely
-  static IconData getIcon(String? type) {
-    if (type == null) return taskTypeIcons['general']!;
-    return taskTypeIcons[type.toLowerCase()] ?? taskTypeIcons['general']!;
-  }
-
-  /// Get an alphabetized list of task type keys
-  static List<String> getSortedTypes() {
-    final keys = taskTypeColors.keys.toList();
-    keys.sort();
-    return keys;
-  }
-}
+import 'widgets/task_form_dialog.dart';
 
 /// =============================
 /// Calendar Assistant Screen
+/// - Displays tasks in a calendar view
+/// - Supports filtering by type and patient
+/// - Integrates with TaskFormDialog to add/edit tasks
 /// =============================
 class CalendarAssistantScreen extends StatefulWidget {
   const CalendarAssistantScreen({super.key});
@@ -417,6 +49,8 @@ class _CalendarAssistantScreenState extends State<CalendarAssistantScreen> {
     equals: isSameDay,
     hashCode: (date) => date.day * 1000000 + date.month * 10000 + date.year,
   );
+
+  /// Apply current filters (task types, patients) to the full task set
   void _applyFilters() {
     setState(() {
       _filteredTasks = {};
@@ -602,7 +236,9 @@ class _CalendarAssistantScreenState extends State<CalendarAssistantScreen> {
     return tasks;
   }
 
-  /// Build fucntion to actual make the Calendar Assistant
+  /// Main widget tree for the screen
+  /// - Shows loading spinner while fetching
+  /// - Renders filter panel, calendar, legend, and task list
   @override
   Widget build(BuildContext context) {
     if (isLoading) {
@@ -799,7 +435,9 @@ class _CalendarAssistantScreenState extends State<CalendarAssistantScreen> {
                 ),
               ),
 
-              // Calendar
+              // ==========================
+              // CALENDAR
+              // ==========================
               TableCalendar<Task>(
                 firstDay: DateTime.utc(2020, 1, 1),
                 lastDay: DateTime.utc(2030, 12, 31),
@@ -838,7 +476,9 @@ class _CalendarAssistantScreenState extends State<CalendarAssistantScreen> {
               ),
               const SizedBox(height: 16),
 
-              // Legend
+              // ==========================
+              // LEGEND (task type dots)
+              // ==========================
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: Wrap(
@@ -855,7 +495,9 @@ class _CalendarAssistantScreenState extends State<CalendarAssistantScreen> {
               ),
               const SizedBox(height: 16),
 
-              // Task list for selected day
+              // ==========================
+              // TASK LIST (for selected day)
+              // ==========================
               if (_selectedDay != null) ...[
                 Builder(
                   builder: (_) {
@@ -952,7 +594,16 @@ class _CalendarAssistantScreenState extends State<CalendarAssistantScreen> {
     );
   }
 
-  /// This fucntion is used to add a task to the CareConnect system
+  // ==========================
+  // TASK CRUD HANDLERS
+  // ==========================
+
+  /// Add a new task to the CareConnect system
+  /// - Opens the [TaskFormDialog] for user input
+  /// - Preloads patient list if caregiver
+  /// - Submits new task to backend via [ApiService.createTaskV2]
+  /// - On success: refreshes tasks from DB and shows confirmation
+  /// - On failure: shows error snackbar
   Future<void> _addTask() async {
     final user = Provider.of<UserProvider>(context, listen: false).user;
     if (user == null) return;
@@ -1005,7 +656,12 @@ class _CalendarAssistantScreenState extends State<CalendarAssistantScreen> {
     }
   }
 
-  /// This function is to edit tasks in the CareConnect system displayed by the Calendar Assistant
+  /// Edit an existing task in the CareConnect system
+  /// - Refreshes the latest version of the task from backend
+  /// - Resolves series anchor date if task is part of a recurrence
+  /// - Opens the [TaskFormDialog] for editing
+  /// - Submits updates via [ApiService.editTaskV2]
+  /// - Supports updating a single task or entire series
   Future<void> _editTask(Task task) async {
     final user = Provider.of<UserProvider>(context, listen: false).user;
     if (user == null) return;
@@ -1108,7 +764,11 @@ class _CalendarAssistantScreenState extends State<CalendarAssistantScreen> {
     }
   }
 
-  /// This function is to remove tasks in the CareConnect system displayed by the Calendar Assistant
+  /// Remove a task from the CareConnect system
+  /// - Prompts user for confirmation
+  /// - If task is part of a recurrence, offers option to delete entire series
+  /// - Calls [ApiService.deleteTaskV2] with appropriate flag
+  /// - On success: reloads tasks and shows confirmation
   Future<void> _removeTask(Task task) async {
     bool applyToSeries = false;
 
@@ -1198,899 +858,5 @@ class _CalendarAssistantScreenState extends State<CalendarAssistantScreen> {
         context,
       ).showSnackBar(SnackBar(content: Text("Error deleting task: $e")));
     }
-  }
-}
-
-// =============================
-// TaskFormDialog.dart
-// =============================
-/// Since Add and Edit use the same form for tasks that logic is broken up here into two forms one being the TaskFormDialog and the other being the RecurrenceForm.
-/// This allows functionality of the forms to be chared across the code.
-class TaskFormDialog extends StatefulWidget {
-  final Task? initialTask;
-  final bool isCaregiver;
-  final List<Map<String, dynamic>> patients;
-  final int? defaultPatientId;
-  final DateTime? initialDate;
-  final DateTime? seriesAnchorDate;
-
-  const TaskFormDialog({
-    super.key,
-    this.initialTask,
-    required this.isCaregiver,
-    required this.patients,
-    this.defaultPatientId,
-    this.initialDate,
-    this.seriesAnchorDate,
-  });
-
-  @override
-  State<TaskFormDialog> createState() => _TaskFormDialogState();
-}
-
-class _TaskFormDialogState extends State<TaskFormDialog> {
-  late TextEditingController titleController;
-  late TextEditingController descriptionController;
-  TimeOfDay? selectedTime;
-  int? selectedPatientId;
-  String? selectedTaskType;
-  // Pull task types directly here
-  late final List<String> taskTypes = TaskTypeUtils.getSortedTypes();
-
-  // Recurrence state
-  bool isRecurring = false;
-  String? recurrenceType;
-  List<bool>? daysOfWeek;
-  int? interval;
-  int? count;
-  DateTime? startDate;
-  DateTime? endDate;
-  int? dayOfMonth;
-  bool applyToSeries = false;
-
-  String? selectedReminder = "None";
-  final List<String> reminderOptions = [
-    "None",
-    "5 minutes before",
-    "15 minutes before",
-    "30 minutes before",
-    "1 hour before",
-    "1 day before",
-  ];
-  static const String _keepExistingCustom = "Keep existing (custom)";
-
-  @override
-  void initState() {
-    super.initState();
-
-    final t = widget.initialTask;
-
-    titleController = TextEditingController(text: t?.name ?? '');
-    descriptionController = TextEditingController(text: t?.description ?? '');
-    selectedTime = t?.timeOfDay;
-    selectedPatientId = widget.defaultPatientId ?? t?.userId;
-    //If editing and taskType is valid, keep it
-    if (t != null &&
-        t.taskType != null &&
-        taskTypes.contains(t.taskType!.toLowerCase())) {
-      selectedTaskType = t.taskType!.toLowerCase();
-    } else {
-      // Otherwise default to "general"
-      selectedTaskType = taskTypes.contains("general")
-          ? "general"
-          : taskTypes.first;
-    }
-    // ensure Save button re-checks when text changes
-    titleController.addListener(() => setState(() {}));
-
-    if (t != null) {
-      recurrenceType = _inferRecurrenceTypeFromTask(t);
-      isRecurring = recurrenceType != null;
-      daysOfWeek = t.daysOfWeek;
-      interval = t.interval;
-      count = t.count;
-      startDate = t.date;
-      final anchorStart = TaskUtils.normalizeDate(
-        widget.seriesAnchorDate ?? t.date,
-      );
-
-      if (recurrenceType == "Monthly") {
-        dayOfMonth = t.date.day;
-      }
-      // compute end date using helper
-      if (count != null &&
-          count! > 0 &&
-          startDate != null &&
-          recurrenceType != null) {
-        final iv = (interval == null || interval! < 1) ? 1 : interval!;
-        endDate = RecurrenceUtils.calculateEndDate(
-          startDate: anchorStart,
-          frequency: recurrenceType!,
-          interval: iv,
-          count: count!,
-          daysOfWeek: daysOfWeek,
-        );
-      }
-    } else {
-      // Use initialDate from the calendar if passed, otherwise fallback to now
-      startDate = TaskUtils.normalizeDate(widget.initialDate ?? DateTime.now());
-    }
-    _prefillReminderFromExistingIfEditing();
-  }
-
-  // ===== Reminder helpers =====
-  Duration? _offsetForLabel(String? label) {
-    switch (label) {
-      case "5 minutes before":
-        return const Duration(minutes: 5);
-      case "15 minutes before":
-        return const Duration(minutes: 15);
-      case "30 minutes before":
-        return const Duration(minutes: 30);
-      case "1 hour before":
-        return const Duration(hours: 1);
-      case "1 day before":
-        return const Duration(days: 1);
-      default:
-        return null;
-    }
-  }
-
-  String? _labelForOffset(Duration diff) {
-    const tol = Duration(minutes: 2);
-    final candidates = <String, Duration>{
-      "5 minutes before": const Duration(minutes: 5),
-      "15 minutes before": const Duration(minutes: 15),
-      "30 minutes before": const Duration(minutes: 30),
-      "1 hour before": const Duration(hours: 1),
-      "1 day before": const Duration(days: 1),
-    };
-
-    for (final e in candidates.entries) {
-      if ((e.value - diff).abs() <= tol) return e.key;
-    }
-    return null;
-  }
-
-  void _prefillReminderFromExistingIfEditing() {
-    final t = widget.initialTask;
-    if (t == null || t.notifications == null || t.notifications!.isEmpty) {
-      selectedReminder = "None";
-      return;
-    }
-
-    // Use first (earliest) notification
-    final notif =
-        (t.notifications!..sort((a, b) {
-              final da =
-                  DateTime.tryParse(a.scheduledTime.toIso8601String()) ??
-                  DateTime(1900);
-              final db =
-                  DateTime.tryParse(b.scheduledTime.toIso8601String()) ??
-                  DateTime(1900);
-              return da.compareTo(db);
-            }))
-            .first;
-
-    if (startDate == null) {
-      selectedReminder = "None";
-      return;
-    }
-
-    final taskDateTime = DateTime(
-      startDate!.year,
-      startDate!.month,
-      startDate!.day,
-      selectedTime?.hour ?? 0,
-      selectedTime?.minute ?? 0,
-    );
-
-    final scheduled = DateTime.tryParse(notif.scheduledTime.toIso8601String());
-    if (scheduled == null) {
-      selectedReminder = "None";
-      return;
-    }
-
-    final diff = taskDateTime.difference(scheduled);
-
-    // Try to match to a standard label
-    final label = _labelForOffset(diff);
-    selectedReminder = label ?? "None";
-  }
-
-  String? _inferRecurrenceTypeFromTask(Task t) {
-    if (t.daysOfWeek?.any((d) => d) ?? false) return "Weekly";
-    switch (t.frequency?.toLowerCase()) {
-      case "daily":
-        return "Daily";
-      case "weekly":
-        return "Weekly";
-      case "monthly":
-        return "Monthly";
-      case "yearly":
-        return "Yearly";
-      default:
-        return null;
-    }
-  }
-
-  /// validation logic of the form's save button
-  bool get canSave {
-    // Must have a name
-    if (titleController.text.trim().isEmpty) return false;
-
-    if (isRecurring) {
-      // Must have a recurrence type
-      if (recurrenceType == null || recurrenceType!.isEmpty) return false;
-      // Weekly: require at least one day
-      if (recurrenceType!.toLowerCase() == "weekly") {
-        if (daysOfWeek == null || !daysOfWeek!.any((d) => d)) {
-          return false;
-        }
-      }
-      // Must have an end condition (endDate or count)
-      if ((endDate == null || endDate!.isBefore(startDate ?? DateTime.now())) &&
-          (count == null || count! <= 0)) {
-        return false;
-      }
-    }
-
-    return true;
-  }
-
-  ///Building the task form based on Add or Edit function
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      title: Text(widget.initialTask == null ? "Add Task" : "Edit Task"),
-      content: SingleChildScrollView(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            // Title
-            TextFormField(
-              controller: titleController,
-              decoration: const InputDecoration(
-                labelText: "Task Title",
-                border: OutlineInputBorder(),
-                floatingLabelBehavior: FloatingLabelBehavior.always,
-              ),
-            ),
-            const SizedBox(height: 12),
-
-            // Description
-            TextFormField(
-              controller: descriptionController,
-              maxLines: 3,
-              decoration: const InputDecoration(
-                labelText: "Description",
-                border: OutlineInputBorder(),
-                floatingLabelBehavior: FloatingLabelBehavior.always,
-              ),
-            ),
-            const SizedBox(height: 12),
-            DropdownButtonFormField<String>(
-              decoration: const InputDecoration(
-                labelText: "Task Type",
-                border: OutlineInputBorder(),
-              ),
-              initialValue:
-                  selectedTaskType != null &&
-                      ![
-                        "daily",
-                        "weekly",
-                        "monthly",
-                        "yearly",
-                      ].contains(selectedTaskType!.toLowerCase())
-                  ? selectedTaskType
-                  : null,
-              items: taskTypes
-                  .map(
-                    (type) => DropdownMenuItem(
-                      value: type,
-                      child: Text(type[0].toUpperCase() + type.substring(1)),
-                    ),
-                  )
-                  .toList(),
-              onChanged: (val) => setState(() => selectedTaskType = val),
-            ),
-
-            const SizedBox(height: 16),
-            // Time picker
-            Row(
-              children: [
-                const Text("Time: "),
-                Text(
-                  selectedTime != null
-                      ? "${selectedTime!.hour.toString().padLeft(2, '0')}:${selectedTime!.minute.toString().padLeft(2, '0')}"
-                      : "Not set",
-                ),
-                const Spacer(),
-                ElevatedButton(
-                  onPressed: () async {
-                    final picked = await showTimePicker(
-                      context: context,
-                      initialTime: selectedTime ?? TimeOfDay.now(),
-                    );
-                    if (picked != null) setState(() => selectedTime = picked);
-                  },
-                  child: const Text("Pick Time"),
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-
-            /// If caregiver then allow assignment to patient. If patient always assigned to themseleves.
-            // Caregiver dropdown
-            if (widget.isCaregiver)
-              DropdownButtonFormField<int>(
-                decoration: const InputDecoration(
-                  labelText: "Assign to Patient",
-                  border: OutlineInputBorder(),
-                  floatingLabelBehavior: FloatingLabelBehavior.always,
-                ),
-                initialValue: selectedPatientId,
-                items:
-                    (widget.patients..sort((a, b) {
-                          final aName =
-                              "${a['patient']?['firstName'] ?? ''} ${a['patient']?['lastName'] ?? ''}";
-                          final bName =
-                              "${b['patient']?['firstName'] ?? ''} ${b['patient']?['lastName'] ?? ''}";
-                          return aName.compareTo(bName);
-                        }))
-                        .map(
-                          (p) => DropdownMenuItem<int>(
-                            value: p['patient']?['id'],
-                            child: Text(
-                              "${p['patient']?['firstName']} ${p['patient']?['lastName']}",
-                            ),
-                          ),
-                        )
-                        .toList(),
-                onChanged: (val) => setState(() => selectedPatientId = val),
-              ),
-            const SizedBox(height: 12),
-
-            // Reminder dropdown
-            DropdownButtonFormField<String>(
-              decoration: const InputDecoration(
-                labelText: "Reminder Notification",
-                border: OutlineInputBorder(),
-              ),
-              initialValue: selectedReminder,
-              items: reminderOptions.map((opt) {
-                return DropdownMenuItem(value: opt, child: Text(opt));
-              }).toList(),
-              onChanged: (val) {
-                setState(() => selectedReminder = val);
-              },
-            ),
-
-            const SizedBox(height: 16),
-            // Recurrence form
-            RecurrenceForm(
-              initialIsRecurring: isRecurring,
-              initialRecurrenceType: recurrenceType,
-              initialDaysOfWeek: daysOfWeek,
-              initialInterval: interval,
-              initialCount: count,
-              initialStartDate: startDate,
-              initialEndDate: endDate,
-              initialDayOfMonth: dayOfMonth,
-              showApplyToSeries: widget.initialTask != null,
-              anchorStartDate: widget.seriesAnchorDate,
-              onChanged:
-                  ({
-                    bool? isRecurring,
-                    String? recurrenceType,
-                    List<bool>? daysOfWeek,
-                    int? interval,
-                    int? count,
-                    DateTime? startDate,
-                    DateTime? endDate,
-                    int? dayOfMonth,
-                    bool? applyToSeries,
-                  }) {
-                    setState(() {
-                      this.isRecurring = isRecurring ?? false;
-                      this.recurrenceType = recurrenceType;
-                      this.daysOfWeek = daysOfWeek;
-                      this.interval = interval;
-                      this.count = count;
-                      this.startDate = startDate ?? this.startDate;
-                      this.endDate = endDate;
-                      this.dayOfMonth = dayOfMonth;
-                      this.applyToSeries = applyToSeries ?? this.applyToSeries;
-                    });
-                  },
-            ),
-          ],
-        ),
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.pop(context, null),
-          child: const Text("Cancel"),
-        ),
-        ElevatedButton(
-          onPressed: canSave
-              ? () {
-                  final rawTask = Task(
-                    id: widget.initialTask?.id,
-                    name: titleController.text,
-                    description: descriptionController.text,
-                    date: startDate ?? DateTime.now(),
-                    timeOfDay: selectedTime,
-                    userId: selectedPatientId,
-                    isComplete: widget.initialTask?.isComplete ?? false,
-                    notifications: widget.initialTask?.notifications,
-                    frequency: recurrenceType,
-                    interval: interval,
-                    count: count,
-                    daysOfWeek: daysOfWeek,
-                    taskType:
-                        (selectedTaskType ??
-                                widget.initialTask?.taskType ??
-                                "general")
-                            .toLowerCase(),
-                  );
-
-                  // Set notifications
-                  List<ScheduledNotification>? notificationsToSave;
-                  if (selectedReminder == "None") {
-                    notificationsToSave = [];
-                  } else if (selectedReminder == _keepExistingCustom) {
-                    notificationsToSave =
-                        widget.initialTask?.notifications ?? [];
-                  } else {
-                    final off = _offsetForLabel(selectedReminder);
-                    final base = DateTime(
-                      rawTask.date.year,
-                      rawTask.date.month,
-                      rawTask.date.day,
-                      rawTask.timeOfDay?.hour ?? 0,
-                      rawTask.timeOfDay?.minute ?? 0,
-                    );
-                    final reminderTime = off == null
-                        ? base
-                        : base.subtract(off);
-                    notificationsToSave = [
-                      ScheduledNotification(
-                        scheduledTime: reminderTime,
-                        title: "Reminder: ${rawTask.name}",
-                        body: rawTask.description.isNotEmpty
-                            ? rawTask.description
-                            : "Don't forget this task.",
-                        notificationType: "TASK_REMINDER",
-                        receiverId: rawTask.userId!,
-                        status: "PENDING",
-                      ),
-                    ];
-                  }
-                  rawTask.notifications = notificationsToSave;
-
-                  final finalTask = RecurrenceUtils.buildTask(
-                    baseTask: rawTask,
-                    isRecurring: isRecurring,
-                    recurrenceType: recurrenceType,
-                    daysOfWeek: daysOfWeek,
-                    interval: interval,
-                    count: count,
-                    startDate: rawTask.date,
-                    endDate: endDate,
-                    dayOfMonth: dayOfMonth,
-                  );
-                  Future.microtask(() {
-                    if (!mounted) return;
-                    Navigator.pop(context, {
-                      'task': finalTask,
-                      'applyToSeries': applyToSeries,
-                    });
-                  });
-                }
-              : null, //disabled if invalid
-          child: const Text("Save"),
-        ),
-      ],
-    );
-  }
-}
-
-// =============================
-// RecurrenceForm Widget
-// =============================
-///This widget allows mutiple places in the code to use the reoccurnce form portion, and ressue its functionality needed specifically for reoccurences.
-class RecurrenceForm extends StatefulWidget {
-  final void Function({
-    bool? isRecurring,
-    String? recurrenceType,
-    List<bool>? daysOfWeek,
-    int? interval,
-    int? count,
-    DateTime? startDate,
-    DateTime? endDate,
-    int? dayOfMonth,
-    bool? applyToSeries,
-  })
-  onChanged;
-
-  final bool initialIsRecurring;
-  final String? initialRecurrenceType;
-  final List<bool>? initialDaysOfWeek;
-  final int? initialInterval;
-  final int? initialCount;
-  final DateTime? initialStartDate;
-  final DateTime? initialEndDate;
-  final int? initialDayOfMonth;
-  final bool applyToSeries;
-  final bool initialApplyToSeries;
-  final bool showApplyToSeries;
-  final DateTime? anchorStartDate;
-
-  const RecurrenceForm({
-    super.key,
-    required this.onChanged,
-    this.initialIsRecurring = false,
-    this.initialRecurrenceType,
-    this.initialDaysOfWeek,
-    this.initialInterval,
-    this.initialCount,
-    this.initialStartDate,
-    this.initialEndDate,
-    this.initialDayOfMonth,
-    this.showApplyToSeries = false,
-    this.applyToSeries = false,
-    this.initialApplyToSeries = false,
-    this.anchorStartDate,
-  });
-
-  @override
-  State<RecurrenceForm> createState() => _RecurrenceFormState();
-}
-
-class _RecurrenceFormState extends State<RecurrenceForm> {
-  bool isRecurring = false;
-  String? recurrenceType;
-  List<bool>? daysOfWeek;
-  int? interval;
-  int? count;
-  DateTime? startDate;
-  DateTime? endDate;
-  int? dayOfMonth;
-  bool applyToSeries = false;
-
-  @override
-  void initState() {
-    super.initState();
-    isRecurring = widget.initialIsRecurring;
-    recurrenceType = widget.initialRecurrenceType;
-    daysOfWeek = widget.initialDaysOfWeek ?? List.filled(7, false);
-    interval = widget.initialInterval;
-    count = widget.initialCount;
-    //Normalize incoming dates
-    startDate = widget.initialStartDate != null
-        ? TaskUtils.normalizeDate(widget.initialStartDate!)
-        : null;
-
-    endDate = widget.initialEndDate != null
-        ? TaskUtils.normalizeDate(widget.initialEndDate!)
-        : null;
-    dayOfMonth = widget.initialDayOfMonth;
-    applyToSeries = widget.initialApplyToSeries;
-  }
-
-  //validation flags for inline error messages
-  bool get isMissingType =>
-      isRecurring && (recurrenceType == null || recurrenceType!.isEmpty);
-
-  bool get isWeeklyInvalid =>
-      recurrenceType == "Weekly" && !(daysOfWeek?.any((d) => d) ?? false);
-
-  bool get isMissingEndCondition =>
-      isRecurring &&
-      ((endDate == null ||
-              (startDate != null && endDate!.isBefore(startDate!))) &&
-          (count == null || count! <= 0));
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        CheckboxListTile(
-          title: const Text("Recurring Task"),
-          value: isRecurring,
-          onChanged: (val) {
-            setState(() => isRecurring = val ?? false);
-            widget.onChanged(
-              isRecurring: isRecurring,
-              recurrenceType: recurrenceType,
-              daysOfWeek: daysOfWeek,
-              interval: interval,
-              count: count,
-              startDate: startDate,
-              endDate: endDate,
-              dayOfMonth: dayOfMonth,
-            );
-          },
-        ),
-        if (isRecurring) ...[
-          const SizedBox(height: 8),
-          // Recurrence type dropdown
-          DropdownButtonFormField<String>(
-            decoration: const InputDecoration(
-              labelText: "Recurrence Type",
-              border: OutlineInputBorder(),
-              floatingLabelBehavior: FloatingLabelBehavior.always,
-            ),
-            initialValue: recurrenceType,
-            items: [
-              "Daily",
-              "Weekly",
-              "Monthly",
-              "Yearly",
-            ].map((t) => DropdownMenuItem(value: t, child: Text(t))).toList(),
-            onChanged: (val) {
-              setState(() => recurrenceType = val);
-              widget.onChanged(
-                isRecurring: isRecurring,
-                recurrenceType: recurrenceType,
-                daysOfWeek: daysOfWeek,
-                interval: interval,
-                count: count,
-                startDate: startDate,
-                endDate: endDate,
-                dayOfMonth: dayOfMonth,
-              );
-            },
-          ),
-          if (isMissingType)
-            const Padding(
-              padding: EdgeInsets.only(top: 4.0),
-              child: Text(
-                "Please select a recurrence type",
-                style: TextStyle(color: Colors.red, fontSize: 12),
-              ),
-            ),
-          if (widget.showApplyToSeries)
-            CheckboxListTile(
-              title: const Text("Apply changes to entire series"),
-              value: applyToSeries,
-              onChanged: (val) {
-                setState(() => applyToSeries = val ?? false);
-                widget.onChanged(
-                  isRecurring: isRecurring,
-                  recurrenceType: recurrenceType,
-                  daysOfWeek: daysOfWeek,
-                  interval: interval,
-                  count: count,
-                  startDate: startDate,
-                  endDate: endDate,
-                  dayOfMonth: dayOfMonth,
-                  applyToSeries: applyToSeries,
-                );
-              },
-            ),
-
-          // Weekly days-of-week picker
-          if (recurrenceType == "Weekly") ...[
-            const SizedBox(height: 12),
-            const Text("Select Days of Week"),
-            Wrap(
-              spacing: 8,
-              children: List.generate(7, (i) {
-                const days = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];
-                return FilterChip(
-                  label: Text(days[i]),
-                  selected: daysOfWeek?[i] ?? false,
-                  onSelected: (selected) {
-                    setState(() => daysOfWeek?[i] = selected);
-                    widget.onChanged(
-                      isRecurring: isRecurring,
-                      recurrenceType: recurrenceType,
-                      daysOfWeek: daysOfWeek,
-                      interval: interval,
-                      count: count,
-                      startDate: startDate,
-                      endDate: endDate,
-                      dayOfMonth: dayOfMonth,
-                    );
-                  },
-                );
-              }),
-            ),
-            if (isWeeklyInvalid)
-              const Padding(
-                padding: EdgeInsets.only(top: 4.0),
-                child: Text(
-                  "Please select at least one day of the week",
-                  style: TextStyle(color: Colors.red, fontSize: 12),
-                ),
-              ),
-          ],
-
-          // Monthly → day-of-month picker
-          if (recurrenceType == "Monthly") ...[
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                const Text("Day of Month:"),
-                const SizedBox(width: 12),
-                DropdownButton<int>(
-                  value: dayOfMonth,
-                  hint: const Text("Select Day"),
-                  items: List.generate(31, (i) => i + 1)
-                      .map(
-                        (d) => DropdownMenuItem(
-                          value: d,
-                          child: Text(d.toString()),
-                        ),
-                      )
-                      .toList(),
-                  onChanged: (val) {
-                    setState(() => dayOfMonth = val);
-                    widget.onChanged(
-                      isRecurring: isRecurring,
-                      recurrenceType: recurrenceType,
-                      daysOfWeek: daysOfWeek,
-                      interval: interval,
-                      count: count,
-                      startDate: startDate,
-                      endDate: endDate,
-                      dayOfMonth: dayOfMonth,
-                    );
-                  },
-                ),
-              ],
-            ),
-          ],
-          const SizedBox(height: 16),
-
-          // Start date picker
-          Row(
-            children: [
-              Text(
-                startDate != null
-                    ? "Starts: ${startDate!.toLocal().toString().split(' ')[0]}"
-                    : "No start date set",
-              ),
-              const Spacer(),
-              TextButton(
-                onPressed: () async {
-                  final picked = await showDatePicker(
-                    context: context,
-                    initialDate: startDate ?? DateTime.now(),
-                    firstDate: DateTime(2000),
-                    lastDate: DateTime(DateTime.now().year + 5),
-                  );
-                  if (picked != null) {
-                    setState(() => startDate = TaskUtils.normalizeDate(picked));
-                    widget.onChanged(
-                      isRecurring: isRecurring,
-                      recurrenceType: recurrenceType,
-                      daysOfWeek: daysOfWeek,
-                      interval: interval,
-                      count: count,
-                      startDate: startDate,
-                      endDate: endDate,
-                      dayOfMonth: dayOfMonth,
-                    );
-                  }
-                },
-                child: const Text("Pick Start Date"),
-              ),
-            ],
-          ),
-          const SizedBox(height: 12),
-          // End date picker
-          if (recurrenceType != "Yearly") ...[
-            Row(
-              children: [
-                Text(
-                  endDate != null
-                      ? "Ends: ${endDate!.toLocal().toString().split(' ')[0]}"
-                      : "No end date set",
-                ),
-                const Spacer(),
-                TextButton(
-                  onPressed: () async {
-                    final picked = await showDatePicker(
-                      context: context,
-                      initialDate: endDate ?? startDate ?? DateTime.now(),
-                      firstDate: startDate ?? DateTime.now(),
-                      lastDate: DateTime(DateTime.now().year + 5),
-                    );
-                    if (picked != null) {
-                      setState(() => endDate = TaskUtils.normalizeDate(picked));
-                      if (recurrenceType != null) {
-                        final iv = (interval == null || interval! < 1)
-                            ? 1
-                            : interval!;
-                        // 👇 critical: choose the correct start for count math
-                        final effectiveStart = applyToSeries
-                            ? (widget.anchorStartDate ??
-                                  startDate ??
-                                  DateTime.now())
-                            : (startDate ?? DateTime.now());
-
-                        count = RecurrenceUtils.calculateCount(
-                          startDate: TaskUtils.normalizeDate(effectiveStart),
-                          endDate: endDate!,
-                          frequency: recurrenceType!,
-                          interval: iv,
-                          daysOfWeek: daysOfWeek,
-                        );
-                      }
-                      widget.onChanged(
-                        isRecurring: isRecurring,
-                        recurrenceType: recurrenceType,
-                        daysOfWeek: daysOfWeek,
-                        interval: interval,
-                        count: count,
-                        startDate: startDate,
-                        endDate: endDate,
-                        dayOfMonth: dayOfMonth,
-                      );
-                    }
-                  },
-                  child: const Text("Pick End Date"),
-                ),
-              ],
-            ),
-          ] else if (recurrenceType == "Yearly") ...[
-            const SizedBox(height: 12),
-            Row(
-              children: [
-                const Text("Ends in Year:"),
-                const SizedBox(width: 12),
-                DropdownButton<int>(
-                  value: endDate?.year,
-                  hint: const Text("Select Year"),
-                  items: List.generate(10, (i) => DateTime.now().year + i)
-                      .map(
-                        (y) => DropdownMenuItem(
-                          value: y,
-                          child: Text(y.toString()),
-                        ),
-                      )
-                      .toList(),
-                  onChanged: (val) {
-                    if (val != null && startDate != null) {
-                      setState(() {
-                        endDate = DateTime(
-                          val,
-                          startDate!.month,
-                          startDate!.day,
-                        );
-                        count = (val - startDate!.year) + 1; // auto-calc count
-                      });
-                      widget.onChanged(
-                        isRecurring: isRecurring,
-                        recurrenceType: recurrenceType,
-                        daysOfWeek: daysOfWeek,
-                        interval: interval,
-                        count: count,
-                        startDate: startDate,
-                        endDate: endDate,
-                        dayOfMonth: dayOfMonth,
-                      );
-                    }
-                  },
-                ),
-              ],
-            ),
-          ],
-
-          if (isMissingEndCondition)
-            const Padding(
-              padding: EdgeInsets.only(top: 4.0),
-              child: Text(
-                "Please provide an end date of the series",
-                style: TextStyle(color: Colors.red, fontSize: 12),
-              ),
-            ),
-        ],
-      ],
-    );
   }
 }
